diseasePredictor/mango-disease-predictor/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.10.0
pillow>=9.0.0
numpy>=1.21.0

